Creates an online debit transaction request class
