+++ lib/pagseguro.rb
@@ -52,6 +52,7 @@ require "pagseguro/transaction_request"
 require "pagseguro/transaction_request/response"
 require "pagseguro/transaction_request/serializer"
 require "pagseguro/boleto_transaction_request"
+require "pagseguro/online_debit_transaction_request"
 
 I18n.load_path += Dir[File.expand_path("../../locales/*.yml", __FILE__)]
