+++ spec/pagseguro/online_debit_transaction_request_spec.rb
@@ -1,0 +1,12 @@
+require "spec_helper"
+
+describe PagSeguro::OnlineDebitTransactionRequest do
+  it_ensures_type PagSeguro::Bank, :bank
+
+  it "sets the bank" do
+    bank = PagSeguro::Bank.new
+    payment = PagSeguro::OnlineDebitTransactionRequest.new(bank: bank)
+
+    expect(payment.bank).to eq(bank)
+  end
+end
